Finding roots using newtons method!
I have done a quick coding challenge set to myself to use newtons method along with some basic high school algebra and a small bit of calculus to calculate the x intercepts of any polynomial. I plan on adding support for handling errors such as divide by zero and overflow; However as it stands those shouldn't be a problem as for any polynomial with solutions assuming you do not choose the axis of symmetry it will work as expected.